Search for intra-robot loop cloures before adding the current descriptor to the pool of embeddings
In the original code the current embedding gets added to the pool of embeddings, and then we use the current embedding to search for the most similar one, which obviously returns the current kf, and gets discarded.
Since the current implementation of scancontext only returns the best candidate, this means that no intra-robot lidar loop cloures where detected.